backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-CORS==3.0.10
psycopg2-binary==2.9.5
python-dotenv==1.0.0
SQLAlchemy==2.0.7
Werkzeug==2.2.3
